requirements.txt
python-docx
gradio==4.44.1
openai
graphviz==0.20.1